custom-elements/custom-element-lifecycle/types-of-callbacks/created-callback.html

<!DOCTYPE html>
<html>
<head>
<title>Created callback of a custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="CREATED callback is invoked after custom element instance is created and its definition is registered">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.register('x-a', {prototype: proto});
    var customElement = new GeneratedConstructor();
    assert_equals(callbackCalledCounter, 1, 'Callback created was not called after ' +
            'element instance was created');
    
}, 'Test created callback when custom element is created in JavaScript');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2></x-a-2>';
    assert_equals(callbackCalledCounter, 1, 'Callback created was not called after ' +
            'element instance was created');
    
}, 'Test created callback when custom element is created in HTML');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-a-3', {prototype: proto});
    var customElement = doc.createElement('x-a-3');
    assert_equals(callbackCalledCounter, 1, 'Callback created was not called after ' +
            'element instance was created');
    
}, 'Test created callback when custom element is created by createElement');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalledCounter++;
    }
    doc.body.innerHTML = '<x-a-4></x-a-4>';
    doc.register('x-a-4', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback created was not called after ' +
            'element instance was created');
    
}, 'Test created callback when custom element is created in HTML before ' +
    'registration of a custom element');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    doc.body.innerHTML = '<x-a-5 id="xa"></x-a-5>';
    var customElement = doc.querySelector('#xa');
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalledCounter++;
    }
    customElement.constructor.prototype = proto;
    assert_equals(callbackCalledCounter, 0,
            'Callback created was called when custom element is unresolved');
    
}, 'Test created callback when custom element is unresolved.');


testInIFrame('../../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalledCounter++;
    }
    doc.register('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'Callback created was not called');
    
}, 'Test created callback. Create custom element, then register. ' + 
    'Document has browsing context');


testInIFrame('../../resources/register-and-create-custom-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;

    assert_equals('Created callback called', doc.querySelector('#log').innerText,
        'Callback created was not called')
}, 'Test created callback. Register custom element, then create. ' + 
    'Document has browsing context');

</script>
</body>
</html>
